Separate validators in different files
Related to issue #2
Related to issue #4
