+++ src/Categories/Validators/index.js
@@ -1,13 +1,5 @@
 export { validateCPF } from './validateCPF';
-
-export const required = value => (value ? undefined : 'Campo obrigatório');
-
-export const numeralCharactersQuantity = quantity => value =>
-  (value && value.replace(/[^\d]/g, '').length !== quantity ? `Deve conter ${quantity} dígitos` : undefined);
-
-export const numeralCharactersMaxQuantity = max => value =>
-  (value && value.replace(/[^\d]/g, '').length > max ? `Deve conter no máximo ${max} dígitos` : undefined);
-
-export const numeralCharactersMinQuantity = min => value =>
-  (value && value.replace(/[^\d]/g, '').length < min ? `Deve conter no mínimo ${min} dígitos` : undefined);
-
+export { numeralCharactersMinQuantity } from './numeralCharactersMinQuantity';
+export { numeralCharactersMaxQuantity } from './numeralCharactersMaxQuantity';
+export { numeralCharactersQuantity } from './numeralCharactersQuantity';
+export { required } from './required';
