+++ src/Categories/Validators/numeralCharactersMaxQuantity.js
@@ -1,0 +1,2 @@
+export const numeralCharactersMaxQuantity = max => value =>
+  (value && value.replace(/[^\d]/g, '').length > max ? `Deve conter no máximo ${max} dígitos` : undefined);
