+++ src/Categories/Validators/numeralCharactersMinQuantity.js
@@ -1,0 +1,2 @@
+export const numeralCharactersMinQuantity = min => value =>
+  (value && value.replace(/[^\d]/g, '').length < min ? `Deve conter no mínimo ${min} dígitos` : undefined);
