+++ src/Categories/Validators/numeralCharactersQuantity.js
@@ -1,0 +1,2 @@
+export const numeralCharactersQuantity = quantity => value =>
+  (value && value.replace(/[^\d]/g, '').length !== quantity ? `Deve conter ${quantity} dígitos` : undefined);
