+++ src/Categories/Validators/required.js
@@ -1,0 +1,1 @@
+export const required = value => (value ? undefined : 'Campo obrigatório');
